pyproject.toml
[tool.poetry]
name = "database-tool"
version = "0.1.0"
description = "A FastAPI-based database optimization tool using Anthropic Claude."
authors = ["Prabha <prabakar.s@optisolbusiness.com>"]
readme = "README.md"
packages = [{include = "database_tool"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
langchain-anthropic = "^0.3.1"
python-dotenv = "^1.0.1"
langchain = "^0.3.14"
jinja2 = "^3.1.5"
python-multipart = "^0.0.20"
fpdf = "^1.7.2"
markdown = "^3.7"
openai = "^1.65.2"

[tool.poetry.scripts]
start = "database_tool.main:app"

